Fixes bug in Structure Radiation module
For frequency-dependent operators,
need to calculate the `T_deriv' term.
This requires evaluating <w|Sigma|w>.
This was done using Yk table.
But the Yk table used was not garenteed to me filled.
i.e., when Yk from inside SR is used, Yk is only filled
if we are not using Qk table, or if we need to fill the Qk table.
If the Qk table already exists, we didn't fill Yk.

This is solved by delecating Sigma_wv to the SR class, which
knows which table is filled, and which to use.

Also: was doing dw(2) wrong!